Split logger logic to separate file
